+++ psycop/common/feature_generation/sequences/timeseries_windower/test_timeserieswindower.py
@@ -1,11 +1,13 @@
 import datetime
 import random
-from collections.abc import Sequence
 from typing import TypeVar
 
 import polars as pl
 import pytest
 
+from psycop.common.feature_generation.sequences.timeseries_windower.timeseries_windower import (
+    window_timeseries,
+)
 from psycop.common.feature_generation.sequences.timeseries_windower.types.event_dataframe import (
     EventDataframeBundle,
 )
@@ -16,7 +18,6 @@ from psycop.common.feature_generation.sequences.timeseries_windower.types.predic
 )
 from psycop.common.feature_generation.sequences.timeseries_windower.types.sequence_dataframe import (
     SequenceColumns,
-    SequenceDataframeBundle,
 )
 from psycop.common.test_utils.str_to_df import str_to_pl_df
 
@@ -208,30 +209,3 @@ class TestTimeserieswindower:
         ).unpack()
 
         assert len(result_df.collect()) == 1
-
-
-def window_timeseries(
-    prediction_times_bundle: PredictiontimeDataframeBundle,
-    event_bundles: Sequence[EventDataframeBundle],
-    lookbehind: datetime.timedelta | None = None,
-) -> SequenceDataframeBundle:
-    pred_time_df, pred_time_cols = prediction_times_bundle.unpack()
-    exploded_dfs = []
-
-    for event_bundle in event_bundles:
-        event_df, event_cols = event_bundle.unpack()
-
-        exploded_df = pred_time_df.join(event_df, on=event_cols.entity_id, how="left")
-
-        if lookbehind is not None:
-            exploded_df = exploded_df.filter(
-                pl.col(event_cols.timestamp)
-                > (pl.col(pred_time_cols.timestamp) - lookbehind),
-            )
-
-        exploded_dfs.append(exploded_df)
-
-    return SequenceDataframeBundle(
-        df=pl.concat(exploded_dfs).drop_nulls(),
-        cols=SequenceColumns(),
-    )
